Fix meet time format in group apppointment
